+++ src/main/java/de/pt400c/defaultsettings/gui/ExportSwitchSegment.java
@@ -26,12 +26,12 @@ public class ExportSwitchSegment extends Segment {
 	}
 	
 	@Override
-	public boolean isSelected(double mouseX, double mouseY) {
+	public boolean isSelected(int mouseX, int mouseY) {
 		return super.isSelected(mouseX, mouseY) || distanceBetweenPoints((float) this.getPosX() + (28F * (1 - processFactor)), (float) this.getPosY() + 7, (float) mouseX, (float) mouseY) <= 9.4F;
 	}
 
 	@Override
-	public void render(float mouseX, float mouseY, float partialTicks) {
+	public void render(int mouseX, int mouseY, float partialTicks) {
 		
 		drawString("Export Mode:", (float) this.getPosX() - 77, (float) this.getPosY() + 3, 0xff5d5d5d);
 		
@@ -129,7 +129,7 @@ public class ExportSwitchSegment extends Segment {
 	}
 	
 	@Override
-	public void hoverCheck(float mouseX, float mouseY) {
+	public void hoverCheck(int mouseX, int mouseY) {
 		if(this.isSelected(mouseX, mouseY) && this.hoverMessage != null) {
 			
 			final ArrayList<String> lines = new ArrayList<String>();
@@ -160,7 +160,7 @@ public class ExportSwitchSegment extends Segment {
 	}
 	
 	@Override
-	public boolean mouseClicked(double mouseX, double mouseY, int button) {
+	public boolean mouseClicked(int mouseX, int mouseY, int button) {
 
 		if (this.isSelected(mouseX, mouseY)) {
 			this.grabbed = true;
@@ -173,14 +173,14 @@ public class ExportSwitchSegment extends Segment {
 	}
 
 	@Override
-	public boolean mouseDragged(double mouseX, double mouseY, int button) {
+	public boolean mouseDragged(int mouseX, int mouseY, int button) {
 		if (!this.isSelected(mouseX, mouseY))
 			this.grabbed = false;
 		return super.mouseDragged(mouseX, mouseY, button);
 	}
 
 	@Override
-	public boolean mouseReleased(double mouseX, double mouseY, int button) {
+	public boolean mouseReleased(int mouseX, int mouseY, int button) {
 		if (this.grabbed) {
 			if (this.isSelected(mouseX, mouseY))
 				this.grabbed = false;
@@ -196,5 +196,4 @@ public class ExportSwitchSegment extends Segment {
 		}
 		return super.mouseReleased(mouseX, mouseY, button);
 	}
-
 }
