+++ src/main/java/de/pt400c/defaultsettings/gui/ScrollableSegment.java
@@ -34,7 +34,7 @@ public class ScrollableSegment extends Segment {
 	private final ScrollbarSegment scrollBar;
 	private final ButtonBulkActionSegment bulkAction;
 	public final SearchbarSegment searchbar;
-	private double distanceY = 0;
+	private float distanceY = 0;
 	private int maxSize = 0;
 	private final byte id;
 	private float velocity = 0;
@@ -117,7 +117,7 @@ public class ScrollableSegment extends Segment {
 	}
 	
 	@Override
-	public void hoverCheck(float mouseX, float mouseY) {
+	public void hoverCheck(int mouseX, int mouseY) {
 		
 		final float offX = (float) (this.getWidth() + 58);
 		final float offY = 53;
@@ -143,7 +143,7 @@ public class ScrollableSegment extends Segment {
 	}
 	
 	@Override
-	public boolean mouseScrolled(double p_mouseScrolled_1_) {
+	public boolean mouseScrolled(float p_mouseScrolled_1_) {
 		this.maxSize = 18 + 20 * (this.list.size() - 1);
 		if(!this.invisible)
 			this.velocity += p_mouseScrolled_1_;
@@ -156,7 +156,7 @@ public class ScrollableSegment extends Segment {
 	}
 
 	@Override
-	public void render(float mouseX, float mouseY, float partialTicks) {
+	public void render(int mouseX, int mouseY, float partialTicks) {
 
 		this.add += (int) (this.velocity);
 		
@@ -288,7 +288,7 @@ public class ScrollableSegment extends Segment {
 			distanceY = 0;
 		}
 
-		final double scaleFactor = MC.mainWindow.getGuiScaleFactor();
+		final float scaleFactor = (float) MC.mainWindow.getGuiScaleFactor();
 
 		glPushMatrix();
 		glEnable(GL_SCISSOR_TEST);
@@ -367,7 +367,7 @@ public class ScrollableSegment extends Segment {
     }
 	
 	@Override
-	public boolean mouseClicked(double mouseX, double mouseY, int button) {
+	public boolean mouseClicked(int mouseX, int mouseY, int button) {
 		if (this.isSelected(mouseX, mouseY)) {
 			((DefaultSettingsGUI) this.gui).resetSelected();
 			for (int i = 0; i < this.list.size(); i++) {
@@ -396,7 +396,7 @@ public class ScrollableSegment extends Segment {
 	}
 
 	@Override
-	public boolean mouseReleased(double mouseX, double mouseY, int button) {
+	public boolean mouseReleased(int mouseX, int mouseY, int button) {
 		this.grabbed = false;
 
 		for (int i = 0; i < this.list.size(); i++) {
@@ -421,7 +421,7 @@ public class ScrollableSegment extends Segment {
 	}
 	
 	@Override
-	public boolean mouseDragged(double mouseX, double mouseY, int button) {
+	public boolean mouseDragged(int mouseX, int mouseY, int button) {
 
 		for (int i = 0; i < this.list.size(); i++) {
 			int yOffTemp = 18 + 20 * i + add;
@@ -480,7 +480,7 @@ class SettingsButtonSegment extends Segment {
 	}
 	
 	@Override
-	public void hoverCheck(float mouseX, float mouseY) {
+	public void hoverCheck(int mouseX, int mouseY) {
 
 		if(this.isSelectedMark(mouseX, mouseY) && this.mark) {
 			ArrayList<String> lines = new ArrayList<String>();
@@ -512,7 +512,7 @@ class SettingsButtonSegment extends Segment {
 	}
 	
 	@Override
-	public void customRender(float mouseX, float mouseY, float customPosX, float customPosY, float partialTicks) {
+	public void customRender(int mouseX, int mouseY, float customPosX, float customPosY, float partialTicks) {
 		if(this.active && this.timer <= Math.PI)
 			timer += 0.3;
 		else if(!this.active && this.timer > 0)
@@ -577,20 +577,20 @@ class SettingsButtonSegment extends Segment {
 	}
 	
 	@Override
-	public boolean isSelected(double mouseX, double mouseY) {
-		final double tempX = this.getPosX() + this.offX; 
-		final double tempY = this.getPosY() + this.offY; 
+	public boolean isSelected(int mouseX, int mouseY) {
+		final float tempX = this.getPosX() + this.offX; 
+		final float tempY = this.getPosY() + this.offY; 
 		return (((GuiConfig) this.gui).popupField == null || this.getIsPopupSegment()) && mouseX >= tempX - 8 && mouseY >= tempY - 4 && mouseX < tempX + this.getWidth() + 2 && mouseY < tempY + this.getHeight() + 4;
 	}
 	
-	public boolean isSelectedMark(double mouseX, double mouseY) {
-		final double tempX = this.getPosX() + this.offX; 
-		final double tempY = this.getPosY() + this.offY; 
+	public boolean isSelectedMark(int mouseX, int mouseY) {
+		final float tempX = this.getPosX() + this.offX; 
+		final float tempY = this.getPosY() + this.offY; 
 		return (((GuiConfig) this.gui).popupField == null || this.getIsPopupSegment()) && mouseX >= tempX - 24 && mouseY >= tempY - 4 && mouseX < tempX + this.getWidth() - 14 && mouseY < tempY + this.getHeight() + 4;
 	}
 
 	@Override
-	public boolean mouseClicked(double mouseX, double mouseY, int button) {
+	public boolean mouseClicked(int mouseX, int mouseY, int button) {
 
 		if (this.isSelected(mouseX, mouseY)) {
 			this.grabbed = true;
@@ -602,7 +602,7 @@ class SettingsButtonSegment extends Segment {
 	}
 
 	@Override
-	public boolean mouseDragged(double mouseX, double mouseY, int button) {
+	public boolean mouseDragged(int mouseX, int mouseY, int button) {
 		if (!this.isSelected(mouseX, mouseY)) {
 			this.grabbed = false;
 			return false;
@@ -611,7 +611,7 @@ class SettingsButtonSegment extends Segment {
 	}
 	
 	@Override
-	public boolean mouseReleased(double mouseX, double mouseY, int button) {
+	public boolean mouseReleased(int mouseX, int mouseY, int button) {
 
 		if (this.grabbed) {
 			if (this.isSelected(mouseX, mouseY))
@@ -651,7 +651,7 @@ class SettingsButtonSegment extends Segment {
 	}
 
 	@Override
-	public void render(float mouseX, float mouseY, float partialTicks) {
+	public void render(int mouseX, int mouseY, float partialTicks) {
 
 	}
 
@@ -680,14 +680,14 @@ class PopupCheckboxSegment extends Segment {
 	}
 	
 	@Override
-	public boolean isSelected(double mouseX, double mouseY) {
-		double tempX = this.getPosX(); 
-		double tempY = this.getPosY(); 
+	public boolean isSelected(int mouseX, int mouseY) {
+		float tempX = this.getPosX(); 
+		float tempY = this.getPosY(); 
 		return (((GuiConfig) this.gui).popupField == null || this.getIsPopupSegment()) && mouseX >= tempX - 4 && mouseY >= tempY - 4 && mouseX < tempX + this.getWidth() + 4&& mouseY < tempY + this.getHeight() + 4;
 	}
 
 	@Override
-	public boolean mouseClicked(double mouseX, double mouseY, int button) {
+	public boolean mouseClicked(int mouseX, int mouseY, int button) {
 
 		if (this.isSelected(mouseX, mouseY)) {
 			this.grabbed = true;
@@ -698,7 +698,7 @@ class PopupCheckboxSegment extends Segment {
 	}
 
 	@Override
-	public boolean mouseDragged(double mouseX, double mouseY, int button) {
+	public boolean mouseDragged(int mouseX, int mouseY, int button) {
 		if (!this.isSelected(mouseX, mouseY)) {
 			this.grabbed = false;
 			return false;
@@ -707,7 +707,7 @@ class PopupCheckboxSegment extends Segment {
 	}
 
 	@Override
-	public boolean mouseReleased(double mouseX, double mouseY, int button) {
+	public boolean mouseReleased(int mouseX, int mouseY, int button) {
 
 		if (this.isSelected(mouseX, mouseY)) {
 			this.grabbed = false;
@@ -889,7 +889,7 @@ class PopupCheckboxSegment extends Segment {
 	}
 
 	@Override
-	public void render(float mouseX, float mouseY, float partialTicks) {
+	public void render(int mouseX, int mouseY, float partialTicks) {
 
 		if (active) {
 
@@ -1001,5 +1001,4 @@ class PopupCheckboxSegment extends Segment {
 		glEnable(GL_TEXTURE_2D);
 		
 	}
-
 }
